10.Governing_AI_Agents/hands-on-ai-governing-ai-agents-in-practice/requirements.txt
openai
python-dotenv
openai-agents
pydantic